# AI_API/docs/openapi.yaml
{"openapi":"3.1.0","info":{"title":"API IA - Prédictions Pandémiques","description":"API spécialisée dans les prédictions ML pour COVID et MPOX","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"Point d'entrée de l'API IA","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Vérification de l'état de santé de l'API","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/test-connection":{"get":{"summary":"Test Express Connection","description":"Test spécifique de la connexion avec l'API Express.js","operationId":"test_express_connection_test_connection_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/mortality/predict":{"get":{"tags":["Mortality Predictions"],"summary":"Predict Mortality","description":"Prédit le taux de mortalité pour les prochains jours à partir de la date de référence (ou aujourd'hui).\nUtilise toutes les données historiques disponibles (de 2020-01-01 à la date de référence) pour entraîner le modèle.","operationId":"predict_mortality_api_mortality_predict_get","parameters":[{"name":"pays","in":"query","required":true,"schema":{"type":"string","description":"Nom du pays ou code ISO","title":"Pays"},"description":"Nom du pays ou code ISO"},{"name":"source","in":"query","required":true,"schema":{"type":"string","description":"Source des données (covid, mpox)","title":"Source"},"description":"Source des données (covid, mpox)"},{"name":"horizon","in":"query","required":false,"schema":{"type":"integer","description":"Horizon de prédiction en jours (défaut: 7)","default":7,"title":"Horizon"},"description":"Horizon de prédiction en jours (défaut: 7)"},{"name":"reference_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Date de référence pour les prédictions (YYYY-MM-DD). Défaut = aujourd'hui","title":"Reference Date"},"description":"Date de référence pour les prédictions (YYYY-MM-DD). Défaut = aujourd'hui"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Predict Mortality Api Mortality Predict Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/mortality/model-info":{"get":{"tags":["Mortality Predictions"],"summary":"Get Mortality Model Info","description":"Retourne les informations sur le modèle de mortalité","operationId":"get_mortality_model_info_api_mortality_model_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Mortality Model Info Api Mortality Model Info Get"}}}}}}},"/api/mortality/data-quality":{"get":{"tags":["Mortality Predictions"],"summary":"Check Data Quality","description":"Vérifie la qualité des données pour la prédiction de mortalité","operationId":"check_data_quality_api_mortality_data_quality_get","parameters":[{"name":"pays","in":"query","required":true,"schema":{"type":"string","description":"Nom du pays ou code ISO","title":"Pays"},"description":"Nom du pays ou code ISO"},{"name":"source","in":"query","required":true,"schema":{"type":"string","description":"Source des données (covid, mpox)","title":"Source"},"description":"Source des données (covid, mpox)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Check Data Quality Api Mortality Data Quality Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/mortality/test-date-configurations":{"get":{"tags":["Mortality Predictions"],"summary":"Test Date Configurations","description":"Teste différentes configurations de dates pour les prédictions\nRetourne les prédictions avec différentes dates de référence pour comparaison","operationId":"test_date_configurations_api_mortality_test_date_configurations_get","parameters":[{"name":"pays","in":"query","required":true,"schema":{"type":"string","description":"Nom du pays ou code ISO","title":"Pays"},"description":"Nom du pays ou code ISO"},{"name":"source","in":"query","required":true,"schema":{"type":"string","description":"Source des données (covid, mpox)","title":"Source"},"description":"Source des données (covid, mpox)"},{"name":"horizon","in":"query","required":false,"schema":{"type":"integer","description":"Horizon de prédiction en jours (défaut: 7)","default":7,"title":"Horizon"},"description":"Horizon de prédiction en jours (défaut: 7)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Test Date Configurations Api Mortality Test Date Configurations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/rt/predict":{"get":{"tags":["Rt Predictions"],"summary":"Predict Rt","description":"Prédit le taux de transmission (Rt) pour les prochains jours à partir de la date de référence (ou aujourd'hui).\nUtilise toutes les données historiques disponibles pour entraîner le modèle LSTM.","operationId":"predict_rt_api_rt_predict_get","parameters":[{"name":"pays","in":"query","required":true,"schema":{"type":"string","description":"Nom du pays ou code ISO","title":"Pays"},"description":"Nom du pays ou code ISO"},{"name":"indicator","in":"query","required":true,"schema":{"type":"string","description":"Indicateur (cases, deaths)","title":"Indicator"},"description":"Indicateur (cases, deaths)"},{"name":"source","in":"query","required":true,"schema":{"type":"string","description":"Source des données (covid, mpox)","title":"Source"},"description":"Source des données (covid, mpox)"},{"name":"horizon","in":"query","required":false,"schema":{"type":"integer","description":"Horizon de prédiction en jours (défaut: 7)","default":7,"title":"Horizon"},"description":"Horizon de prédiction en jours (défaut: 7)"},{"name":"reference_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Date de référence pour les prédictions (YYYY-MM-DD). Défaut = aujourd'hui","title":"Reference Date"},"description":"Date de référence pour les prédictions (YYYY-MM-DD). Défaut = aujourd'hui"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Predict Rt Api Rt Predict Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/rt/model-info":{"get":{"tags":["Rt Predictions"],"summary":"Get Rt Model Info","description":"Retourne les informations sur le modèle Rt (LSTM)","operationId":"get_rt_model_info_api_rt_model_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Rt Model Info Api Rt Model Info Get"}}}}}}},"/api/rt/data-quality":{"get":{"tags":["Rt Predictions"],"summary":"Check Rt Data Quality","description":"Vérifie la qualité des données pour la prédiction Rt","operationId":"check_rt_data_quality_api_rt_data_quality_get","parameters":[{"name":"pays","in":"query","required":true,"schema":{"type":"string","description":"Nom du pays ou code ISO","title":"Pays"},"description":"Nom du pays ou code ISO"},{"name":"indicator","in":"query","required":true,"schema":{"type":"string","description":"Indicateur (cases, deaths)","title":"Indicator"},"description":"Indicateur (cases, deaths)"},{"name":"source","in":"query","required":true,"schema":{"type":"string","description":"Source des données (covid, mpox)","title":"Source"},"description":"Source des données (covid, mpox)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Check Rt Data Quality Api Rt Data Quality Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/spread/predict":{"get":{"tags":["Spread Predictions","Spread Predictions"],"summary":"Clustering géographique des pays (propagation)","description":"Prédit la propagation géographique par clustering KMeans.\n\n- Récupère les séries temporelles de chaque pays via l'API Express.js (/geographic-spread).\n- Applique un clustering KMeans sur les dynamiques temporelles de l'indicateur choisi (cases ou deaths).\n- Retourne les groupes de pays similaires, avec métadonnées et explications.\n\n**Exemples d'usage :**\n- `/api/spread/predict?indicator=cases&source=covid&k=3`\n- `/api/spread/predict?indicator=deaths&source=mpox&k=4`\n\n**Réponse :**\n- `clusters` : liste des groupes, chaque groupe contient les pays (code ISO, nom)\n- `meta` : informations sur le clustering, l'indicateur, la source, le modèle utilisé, etc.","operationId":"predict_spread_api_spread_predict_get","parameters":[{"name":"indicator","in":"query","required":true,"schema":{"type":"string","description":"Indicateur à clusteriser (cases, deaths)","title":"Indicator"},"description":"Indicateur à clusteriser (cases, deaths)","example":"cases"},{"name":"source","in":"query","required":true,"schema":{"type":"string","description":"Source des données (covid, mpox)","title":"Source"},"description":"Source des données (covid, mpox)","example":"covid"},{"name":"k","in":"query","required":false,"schema":{"type":"integer","maximum":20,"minimum":2,"description":"Nombre de clusters à former (défaut: 3)","default":3,"title":"K"},"description":"Nombre de clusters à former (défaut: 3)","example":3}],"responses":{"200":{"description":"Clusters de pays similaires","content":{"application/json":{"schema":{"type":"object","title":"Response Predict Spread Api Spread Predict Get"},"example":{"clusters":[{"cluster":1,"countries":[{"iso_code":"TCD","country":"Chad"},{"iso_code":"KGZ","country":"Kyrgyzstan"},{"iso_code":"PCN","country":"Pitcairn"},{"iso_code":"TKL","country":"Tokelau"},{"iso_code":"TKM","country":"Turkmenistan"},{"iso_code":"TUV","country":"Tuvalu"},{"iso_code":"VAT","country":"Vatican"},{"iso_code":"ASM","country":"American Samoa"},{"iso_code":"ARG","country":"Argentina"},{"iso_code":"ABW","country":"Aruba"},{"iso_code":"BHS","country":"Bahamas"},{"iso_code":"BHR","country":"Bahrain"},{"iso_code":"BRB","country":"Barbados"},{"iso_code":"BTN","country":"Bhutan"},{"iso_code":"BRA","country":"Brazil"},{"iso_code":"VGB","country":"British Virgin Islands"},{"iso_code":"CPV","country":"Cape Verde"},{"iso_code":"NPL","country":"Nepal"},{"iso_code":"RWA","country":"Rwanda"}]},{"cluster":2,"countries":[{"iso_code":"AGO","country":"Angola"},{"iso_code":"BWA","country":"Botswana"},{"iso_code":"BFA","country":"Burkina Faso"},{"iso_code":"FLK","country":"Falkland Islands"},{"iso_code":"PYF","country":"French Polynesia"},{"iso_code":"PSE","country":"Palestine"},{"iso_code":"ALB","country":"Albania"},{"iso_code":"DZA","country":"Algeria"},{"iso_code":"AND","country":"Andorra"},{"iso_code":"AIA","country":"Anguilla"},{"iso_code":"ARM","country":"Armenia"},{"iso_code":"AUT","country":"Austria"},{"iso_code":"AZE","country":"Azerbaijan"},{"iso_code":"BLR","country":"Belarus"},{"iso_code":"BEL","country":"Belgium"},{"iso_code":"BLZ","country":"Belize"},{"iso_code":"BMU","country":"Bermuda"}]},{"cluster":3,"countries":[{"iso_code":"GNB","country":"Guinea-Bissau"},{"iso_code":"HTI","country":"Haiti"},{"iso_code":"KAZ","country":"Kazakhstan"},{"iso_code":"MDG","country":"Madagascar"},{"iso_code":"AFG","country":"Afghanistan"},{"iso_code":"ATG","country":"Antigua and Barbuda"},{"iso_code":"AUS","country":"Australia"},{"iso_code":"BEN","country":"Benin"},{"iso_code":"BOL","country":"Bolivia"},{"iso_code":"BDI","country":"Burundi"},{"iso_code":"KHM","country":"Cambodia"}]}],"meta":{"indicator":"cases","source":"covid","k":3,"count":234}}}}},"400":{"description":"Paramètres manquants ou invalides"},"500":{"description":"Erreur interne du serveur"},"501":{"description":"Clustering déjà effectué côté Express.js"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/spread/model-info":{"get":{"tags":["Spread Predictions"],"summary":"Get Spread Model Info","description":"Retourne les informations sur le modèle de propagation","operationId":"get_spread_model_info_api_spread_model_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Spread Model Info Api Spread Model Info Get"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}